% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Coda.R
\name{CoDa}
\alias{CoDa}
\title{Fit CoDa model}
\usage{
CoDa(dx, x = NULL, y = NULL)
}
\arguments{
\item{dx}{Female death rates matrix with ages as row and time as column}

\item{x}{Vector of input ages (optional)}

\item{y}{Vector of input years (optional)}
}
\value{
An object of class \code{CoDa}
}
\description{
Fit Compositional Data Analysis (CoDa) model for forecasting the life table 
distribution of deaths.
}
\examples{
# Fit CoDa model
fit_CoDa <- CoDa(CoDa.data, x = 0:110, y = 1960:2014)
ls(fit_CoDa)
summary(fit_CoDa)

# Predict life expectancy 20 years in the future using CoDa model
pred_Coda <- predict(fit_CoDa, n = 20)
pred_Coda

}
\references{
\enumerate{
\item{Bergeron-Boucher, M-P., Canudas-Romo, V., Oeppen, J. and Vaupel, W.J. 2017. 
\href{http://doi.org/10.4054/DemRes.2017.37.17}{
Coherent forecasts of mortality with compositional data analysis.}
Demographic Research, Volume 17, Article 17, Pages 527--566.}
\item{Aitchison, J. 1986. 
\href{http://www.leg.ufpr.br/lib/exe/fetch.php/pessoais:abtmartins:a_concise_guide_to_compositional_data_analysis.pdf}{
The Statistical Analysis of Compositional Data.} 
London: Chapman and Hall. 2015.}
}
}
\seealso{
\code{\link{predict.CoDa}}
}
