\name{codamenu}
\alias{codamenu}
\title{Main menu driver for the coda package}

\usage{
codamenu()
}

\description{
\code{codamenu} presents a simple menu-based interface to the functions
in the coda package.  It is designed for users who know nothing about
the R/S language.}

\author{Kate Cowles, Nicky Best, Karen Vines, Martyn Plummer}
\keyword{utilities}
