\name{Cramer}
\alias{pcramer}
\title{The Cramer-von Mises Distribution}
\description{
  Distribution function of the Cramer-von Mises distribution.
}
\usage{
  pcramer(q, eps)
}
\arguments{
  \item{q}{vector of quantiles.}
  \item{eps}{accuracy required}
}
\value{
  \code{pcramer} gives the distribution function,
}
\references{
Anderson TW. and Darling DA. Asymptotic theory of certain `goodness
of fit' criteria based on stochastic processes.  \emph{Ann. Math.
Statist.}, \bold{23}, 192-212 (1952).

Csorgo S. and Faraway, JJ. The exact and asymptotic distributions of
the Cramer-von Mises statistic.  J. Roy. Stat. Soc. (B), \bold{58},
221-234 (1996).
}

\keyword{distribution}

