% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edd}
\alias{edd}
\title{Data - for testing purposes}
\format{An object of class \code{matrix} with 111 rows and 55 columns.}
\source{
Human Mortality Database, \url{http://www.mortality.org}.
}
\usage{
edd
}
\description{
Dataset containing empirical distribution of deaths
for US female population between 1960 and 2014. 
This data is provided for testing purposes only.
Download the actual data free of charge from \url{http://www.mortality.org}.
}
\seealso{
\code{\link{CoDa}}
}
\keyword{datasets}
