% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmdus}
\alias{hmdus}
\title{DATA - for testing purposes}
\format{A list with 4 matrices with 101 rows and 52 columns:
\describe{
  \item{rows}{age interval}
  \item{colums}{years}
}}
\source{
Human Mortality Database, \url{http://www.mortality.org}.
}
\usage{
hmdus
}
\description{
Dataset containing the HMD death rates and death counts 
for US population between 1960 and 2012. 
This data is provided for testing purposes only.
Download the actual data free of charge from \url{http://www.mortality.org}.
}
\keyword{datasets}
