\name{line}
\alias{line}
\alias{line.old}
\title{
Simple linear regression example
}
\usage{
data(line)
}
\description{
    Sample MCMC output from a simple linear regression model given in
    the BUGS manual.

    Dataset line.old contains the same data in an old format used by
    versions of the coda library up to 0.3.
}
\source{
Spiegelhalter, D.J., Thomas, A., Best, N.G. and Gilks, W.R. (1995)
BUGS: Bayesian inference using Gibbs Sampling, Version 0.5, 
MRC Biostatistics Unit, Cambridge.
}
\keyword{datasets}
