\name{line}
\docType{data}
\alias{line}
\title{Simple linear regression example}
\description{
  Sample MCMC output from a simple linear regression model given in
  the BUGS manual.
}
\usage{data(line)}
\format{An \code{mcmc} object}
\source{
  Spiegelhalter, D.J., Thomas, A., Best, N.G. and Gilks, W.R. (1995)
  BUGS: Bayesian inference using Gibbs Sampling, Version 0.5, 
  MRC Biostatistics Unit, Cambridge.
}
\keyword{datasets}
