% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{plot.coda}
\alias{plot.coda}
\title{Plot fitted parameters from a stochastic CoDa mortality model}
\usage{
\method{plot}{coda}(x, type = "l", ...)
}
\arguments{
\item{x}{An object of class \code{"coda"} with the fitted 
parameters of the mortality model.}

\item{type}{Type of plot to be drawn. See \code{\link[graphics]{plot}}.}

\item{...}{Additional arguments to control graphical appearance.}
}
\description{
Plot fitted parameters from a stochastic CoDa mortality model
}
\examples{
# Fit model
M <- coda(CoDa.data, x = 0:110, y = 1960:2014)

# Plot fitted parameters
plot(M)
plot(M, type = "p", pch = 19)
}
