% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Coda.R
\name{summary.CoDa}
\alias{summary.CoDa}
\title{Summary function for a CoDa object}
\usage{
\method{summary}{CoDa}(object, ...)
}
\arguments{
\item{object}{An object of class CoDa}

\item{...}{...}
}
\value{
Summary
}
\description{
Summary function for a CoDa object
}
\keyword{internal}
