\name{time.mcmc}
\alias{time.mcmc}
\alias{start.mcmc}
\alias{end.mcmc}
\alias{frequency.mcmc}
\alias{thin.mcmc}
\alias{time.mcmc.list}
\alias{start.mcmc.list}
\alias{end.mcmc.list}
\alias{thin.mcmc.list}
\title{Time attributes for mcmc objects}

\usage{
time(x)
start(x)
end(x)
thin(x)
}
\description{
   These are methods for mcmc objects for the generic time
   series functions.
}

\seealso{
   \code{\link{time}},
   \code{\link{start}},
   \code{\link{frequency}},
   \code{\link{thin}}.
}
\keyword{ts}
