\name{as.ts.mcmc}
\title{Coerce mcmc object to time series}

\usage{as.ts.mcmc(mcmc.obj)}

\description{
\code{as.ts.mcmc} will coerce an mcmc object to a time series.}

\notes{
This is pretending to be an \code{as.ts} method for mcmc objects. But
\code{as.ts} is not generic.}

\author{Martyn Plummer}

\seealso{
\code{\link{as.ts}}
}
