\name{as.ts.mcmc}
\alias{as.ts.mcmc}
\title{Coerce mcmc object to time series}

\usage{\method{as.ts}{mcmc}(x,\dots)}

\arguments{
   \item{x}{an mcmc object}
   \item{\dots}{unused arguments for compatibility with generic \code{as.ts}}
}

\description{
the \code{as.ts} method for \code{mcmc} objects coerces an mcmc object to a time series.}

\author{Martyn Plummer}

\seealso{
\code{\link{as.ts}}
}
\keyword{ts}
\keyword{utilities}
