\name{as.ts.mcmc}
\alias{as.ts.mcmc}
\title{Coerce mcmc object to time series}

\usage{as.ts.mcmc(x,...)}

\arguments{
   \item{x}{an mcmc object}
   \item{...}{unused arguments for compatibility with generic \code{as.ts}}
}

\description{
\code{as.ts.mcmc} will coerce an mcmc object to a time series.}

\author{Martyn Plummer}

\seealso{
\code{\link{as.ts}}
}
\keyword{ts}
\keyword{utilities}
