% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Coda.R
\name{CoDa}
\alias{CoDa}
\title{Function to fit death rates using the CoDa model}
\usage{
CoDa(dx, x = NULL, y = NULL)
}
\arguments{
\item{dx}{female death rates matrix with ages as row and time as column}

\item{x}{Vector of input ages (optional)}

\item{y}{Vector of input years (optional)}
}
\value{
An object of class \code{CoDa}
}
\description{
Function to fit death rates using the CoDa model
}
\examples{
# Fit CoDa model
fit_CoDa <- CoDa(CoDa.data, x = 0:110, y = 1960:2014)
ls(fit_CoDa)
summary(fit_CoDa)

# Predict life expectancy 20 years in the future using CoDa model
pred_Coda <- predict(fit_CoDa, n = 20)
pred_Coda

}
\seealso{
\code{\link{predict.CoDa}}
}
