% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_coda.R
\name{coda}
\alias{coda}
\title{Fit CoDa Mortality Model}
\usage{
coda(dx, x = NULL, y = NULL)
}
\arguments{
\item{dx}{Matrix containing mortality data (dx) with ages as row and time as column.}

\item{x}{Vector of input ages (optional). Used to label the output objects and plots.}

\item{y}{Vector of input years (optional). Used to label the output objects and plots.}
}
\value{
The output is an object of class \code{"coda"} with the components:

\item{input}{List with arguments provided in input. Saved for convenience.}

\item{coefficients}{Estimated coefficients.}

\item{fitted}{Fitted values of the estimated CoDa model.}

\item{residuals}{Deviance residuals.}

\item{x}{Vector of ages used in the fitting.}

\item{y}{Vector of years used in the fitting.}

\item{call}{The unevaluated expression of the defined coda function.}
}
\description{
Fit Compositional Data (CoDa) model for forecasting the life table 
distribution of deaths. CoDa is a Lee-Carter type model. A key difference 
between the Lee-Carter (1992) method and the Compositional Data (CoDa) model is that the 
former fits and forecasts the death rates (mx) while the latter is based on the 
life table death distribution (dx). See Bergeron-Boucher et al. (2017) for a 
detail description and mathematical formulation.
}
\examples{
# Fit CoDa model
M <- coda(CoDa.data)
summary(M)

# Forecast life expectancy
P <- predict(M, h = 20)

}
\references{
\enumerate{
\item{Bergeron-Boucher M-P., Canudas-Romo V., Oeppen J. and Vaupel W.J. 2017. 
\href{http://doi.org/10.4054/DemRes.2017.37.17}{
Coherent forecasts of mortality with compositional data analysis.}
Demographic Research, Volume 17, Article 17, Pages 527--566.}
\item{Oeppen, J. 2008. \href{https://dugi-doc.udg.edu/handle/10256/742}{
Coherent forecasting of multiple-decrement life tables: 
A test using Japanese cause of death data.} Paper presented at the 
European Population Conference 2008, Barcelona, Spain, July 9-12, 2008.}
\item{Aitchison, J. 1986. 
\href{http://www.leg.ufpr.br/lib/exe/fetch.php/pessoais:abtmartins:a_concise_guide_to_compositional_data_analysis.pdf}{
The Statistical Analysis of Compositional Data.} 
London: Chapman and Hall. 2015.}
\item{Ronald D. Lee and Lawrence R. Carter. 1992. 
\href{http://doi.org/10.1080/01621459.1992.10475265}{Modeling and Forecasting U.S. Mortality}, 
Journal of the American Statistical Association, 87:419, 659--671.}
}
}
\seealso{
\code{\link{predict.coda}}
}
