% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Coda.R
\name{predict.CoDa}
\alias{predict.CoDa}
\title{Function to predict life expectancy using CoDa model}
\usage{
\method{predict}{CoDa}(object, n, ...)
}
\arguments{
\item{object}{CoDa object}

\item{n}{Number of years to be forcast in the future}

\item{...}{...}
}
\value{
Results
}
\description{
Function to predict life expectancy using CoDa model
}
