% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Coda.R
\name{predict.CoDa}
\alias{predict.CoDa}
\title{Predict empirical distribution of deaths using CoDa model}
\usage{
\method{predict}{CoDa}(object, n, order = NULL, include.drift = NULL,
  method = "ML", ci = c(80, 95), jumpchoice = c("actual", "fit"), ...)
}
\arguments{
\item{object}{CoDa object}

\item{n}{Number of years to be forecast in the future}

\item{order}{A specification of the non-seasonal part of the ARIMA model: 
the three components (p, d, q) are the AR order, the degree of differencing, 
and the MA order. If \code{order = NULL}, the ARIMA order will be estimated 
automatically using the KPPS algorithm.}

\item{include.drift}{Logical. Should the ARIMA model include a linear drift term?
If \code{include.drift = NULL}, the model will be estimated automatically.}

\item{method}{Fitting method: maximum likelihood or minimize conditional 
sum-of-squares. Options to use:
conditional-sum-of-squares (\code{"CSS-ML"}), maximum likelihood (\code{"ML"}) 
and \code{"CSS"}.}

\item{ci}{Confidence level for prediction intervals.}

\item{jumpchoice}{Method used for computation of jumpchoice. 
Possibilities: \code{"actual"} (use actual rates from final year) 
and \code{"fit"} (use fitted rates).}

\item{...}{Additional arguments to be passed to \code{\link{Arima}}}
}
\value{
Results
}
\description{
Predict empirical distribution of deaths using CoDa model
}
